pyproject.toml
[tool.poetry]
name = "ha-mqtt-discoverable"
version = "0.7.0"
description = ""
authors = ["Joe Block <jpb@unixorn.net>"]
readme = "README.md"
packages = [{include = "ha_mqtt_discoverable"}]

[tool.poetry.dependencies]
python = ">=3.10.0,<4.0"
pyaml = "^21.10.1"
paho-mqtt = "^1.6.1"
gitlike-commands = "^0.2.1"
thelogrus = "^0.7.0"
pydantic = "^1.10.5"

typer = {version = "^0.7.0", optional = true, extras = ["rich"]}

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.2.2"
flake8 = "^6.0.0"
pre-commit = "^3.1.1"


[tool.poetry.group.test.dependencies]
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
hmd = 'ha_mqtt_discoverable.cli.main:app'

[tool.poetry.extras]
cli = ["typer"]
